requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h93a5062_5
ca-certificates=2023.11.17=hf0a4a13_0
colorama=0.4.6=pyhd8ed1ab_0
freetype=2.12.1=hadb7bae_2
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=20_osxarm64_openblas
libcblas=3.9.0=20_osxarm64_openblas
libcxx=16.0.6=h4653b0c_0
libdeflate=1.19=hb547adb_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_1
libgfortran5=13.2.0=hf226fd6_1
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=20_osxarm64_openblas
libopenblas=0.3.25=openmp_h6c19121_0
libpng=1.6.39=h76d750c_0
libsqlite=3.44.2=h091b4b1_0
libtiff=4.6.0=ha8a6c65_2
libwebp-base=1.3.2=hb547adb_0
libxcb=1.15=hf346824_0
libzlib=1.2.13=h53f4e23_5
llvm-openmp=17.0.6=hcd81f8e_0
ncurses=6.4=h463b476_2
numpy=1.26.2=py310h30ee222_0
openjpeg=2.5.0=h4c1507b_3
openssl=3.2.0=h0d3ecfb_1
pillow=10.1.0=py310hfae7ebd_0
pip=23.3.2=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
python=3.10.10=h3ba56d0_0_cpython
python_abi=3.10=4_cp310
readline=8.2=h92ec313_1
setuptools=68.2.2=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tqdm=4.66.1=pyhd8ed1ab_0
tzdata=2023c=h71feb2d_0
wheel=0.42.0=pyhd8ed1ab_0
xorg-libxau=1.0.11=hb547adb_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
zstd=1.5.5=h4f39d0f_0
